fix(interface): move interface
